meta-luv: Update kernel selftest recipes
Kernel Commit b2d35fa "selftests: add headers_install to lib.mk"
requires every test-directory in tools/testing/selftests which includes
the lib.mk file to have access to the kernel root directory
(top_srcdir):
  +top_srcdir ?= ../../../..
  +include $(top_srcdir)/scripts/subarch.include
  +ARCH           ?= $(SUBARCH)

In LUV, these tests are the efivarfs-test and the pstore-test. Currently
for both these recipes, we simply copy the respective tests to a
directory and build it from there. Since we do not copy the entire kernel,
we see some errors like the following due to missing files:
make: *** No rule to make target `../../../../scripts/subarch.include'. 
Stop.

Instead, we make changes to the existing recipes so that the tests are
built from the kernel source directory itself, without having to copy it
to separate directories.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>